backend/pyproject.toml
[project]
name = "invoice-maid"
version = "0.9.0a1"
description = "AI-powered invoice extraction service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0.30,<3.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "apscheduler>=3.10,<4.0",
    "openai>=1.30",
    "instructor>=1.5,<2.0",
    "pydantic>=2.7",
    "pydantic[email]>=2.7",
    "pydantic-settings>=2.0",
    "imap-tools>=1.7",
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=3.2,<5.0",
    "python-multipart>=0.0.22",
    "aiofiles>=23.0",
    "filetype>=1.2,<2.0",
    "httpx>=0.27",
    "sse-starlette>=1.6",
    "msal>=1.28",
    "msgraph-sdk>=1.0",
    "lxml>=5.0",
    "cryptography>=42.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
full = [
    "sqlite-vec>=0.1.6",
    "pyzbar>=0.1.9",
    "easyofd>=0.5.6",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=5.0",
    "anyio[trio]>=4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
